puntos en cliente model, canje millas

+++ src/AerolineaFrba/Models/Catalogo.cs
@@ -9,8 +9,8 @@ namespace AerolineaFrba.Models {
 
         public int Id { get; set; }
         public string Descripcion { get; set; }
-        public int Costo { get; set; }
-        public int Stock { get; set; }
+        public Decimal Costo { get; set; }
+        public Decimal Stock { get; set; }
 
         public static string TableName = "BIEN_MIGRADO_RAFA.Catalogo";
